# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/user": {
      "post": {
        "summary": "Register",
        "operationId": "register_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/password/reset": {
      "put": {
        "summary": "Reset User Password",
        "operationId": "reset_user_password_password_reset_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get User List",
        "operationId": "get_user_list_users_get",
        "parameters": [
          {
            "name": "exclude_user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude User Id"
            }
          },
          {
            "name": "min_age",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Age"
            }
          },
          {
            "name": "max_age",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age"
            }
          },
          {
            "name": "birthday",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Birthday"
            }
          },
          {
            "name": "min_height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Height"
            }
          },
          {
            "name": "max_height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Height"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GenderEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gender"
            }
          },
          {
            "name": "smoking",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SmokingEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Smoking"
            }
          },
          {
            "name": "alcohol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlcoholEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alcohol"
            }
          },
          {
            "name": "marriage_intention",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MarriageIntentionEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Marriage Intention"
            }
          },
          {
            "name": "meeting_preference",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeetingPreferenceEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Meeting Preference"
            }
          },
          {
            "name": "living_arrangement",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LivingArrangementEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Living Arrangement"
            }
          },
          {
            "name": "education",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EducationEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Education"
            }
          },
          {
            "name": "income",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncomeEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Income"
            }
          },
          {
            "name": "holiday",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/HolidayEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holiday"
            }
          },
          {
            "name": "current_location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current Location Id"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response Get User List Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "summary": "Get User Detail",
        "operationId": "get_user_detail_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}": {
      "delete": {
        "summary": "Delete",
        "operationId": "delete_user__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "summary": "Get Me",
        "operationId": "get_me_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me": {
      "put": {
        "summary": "Update",
        "operationId": "update_users_me_put",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_users_me_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/{user_id}/footprint": {
      "post": {
        "summary": "Create Footprint",
        "operationId": "create_footprint_users__user_id__footprint_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/footprint": {
      "get": {
        "summary": "Get Visited User",
        "operationId": "get_visited_user_users_me_footprint_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlcoholEnum": {
        "type": "string",
        "enum": [
          "よく飲む",
          "飲む",
          "少し飲む",
          "飲まない"
        ],
        "title": "AlcoholEnum"
      },
...
      "EducationEnum": {
        "type": "string",
        "enum": [
          "高校・高校卒",
          "専門学校・専門卒",
          "短大・短大卒",
          "大学",
          "大学院"
        ],
        "title": "EducationEnum"
      },
      "GenderEnum": {
        "type": "string",
        "enum": [
          "女性",
          "男性",
          "答えたくない"
        ],
        "title": "GenderEnum"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HolidayEnum": {
        "type": "string",
        "enum": [
          "平日休み",
          "土日休み",
          "不定期"
        ],
        "title": "HolidayEnum"
      },
      "IncomeEnum": {
        "type": "string",
        "enum": [
          "200万以下",
          "200~400万",
          "400~600万",
          "600~800万",
          "800~1000万",
          "1000万以上"
        ],
        "title": "IncomeEnum"
      },
...
      "LivingArrangementEnum": {
        "type": "string",
        "enum": [
          "一人暮らし",
          "実家",
          "ルームシェア",
          "友人",
          "その他"
        ],
        "title": "LivingArrangementEnum"
      },
      "LocationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "LocationResponse"
      },
...
      "MarriageIntentionEnum": {
        "type": "string",
        "enum": [
          "すぐにでもしたい",
          "2~3年のうちに",
          "良い人がいればしたい",
          "今はしたくない"
        ],
        "title": "MarriageIntentionEnum"
      },
...
      "MeetingPreferenceEnum": {
        "type": "string",
        "enum": [
          "マッチング後、まずは会いたい",
          "気が合えば会いたい",
          "メッセージを重ねてから会いたい"
        ],
        "title": "MeetingPreferenceEnum"
      },
...
      "PasswordReset": {
        "properties": {
          "mail_address": {
            "type": "string",
            "format": "email",
            "title": "Mail Address"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "password_confirm": {
            "type": "string",
            "title": "Password Confirm"
          }
        },
        "type": "object",
        "required": [
          "mail_address",
          "password",
          "password_confirm"
        ],
        "title": "PasswordReset"
      },
...
      "SmokingEnum": {
        "type": "string",
        "enum": [
          "吸う",
          "相手が嫌ならやめる",
          "吸わない"
        ],
        "title": "SmokingEnum"
      },
...
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "title": "Birthday"
          },
          "mail_address": {
            "type": "string",
            "format": "email",
            "title": "Mail Address"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenderEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "smoking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SmokingEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "alcohol": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlcoholEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "marriage_intention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarriageIntentionEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "meeting_preference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeetingPreferenceEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "living_arrangement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LivingArrangementEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "education": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncomeEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "holiday": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HolidayEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "birth_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Location Id"
          },
          "current_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "age",
          "birthday",
          "password"
        ],
        "title": "UserCreate"
      },
...
      "UserImageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_url",
          "sort_order"
        ],
        "title": "UserImageResponse"
      },
      "UserResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mail_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mail Address"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenderEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "smoking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SmokingEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "alcohol": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlcoholEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "marriage_intention": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarriageIntentionEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "meeting_preference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeetingPreferenceEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "living_arrangement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LivingArrangementEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "education": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EducationEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "income": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncomeEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "holiday": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HolidayEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "birth_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Location Id"
          },
          "current_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "current_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/UserImageResponse"
            },
            "type": "array",
            "title": "Images",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}